add some error checks fix some stuff

+++ setup.sh
@@ -4,23 +4,52 @@
 export DOTFILES_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 
 
-# Simlinks
+# Package Managers and packages & Simlinks
 echo "# Symlink Instalation ✏️"
 ## Git
+## TODO: Fix the gitconfig kinda broken
 ln -sfv "$DOTFILES_DIR/shell/git/.gitconfig" ~
 ln -sfv "$DOTFILES_DIR/shell/git/.gitignore_global" ~
 ln -sfv "$DOTFILES_DIR/shell/git/.gitsettings" ~
+
 ## Zsh
-ln -sf "$DOTFILES_DIR/shell/zsh/.zshrc" "$HOME/.zshrc"
+if [ -f "$HOME/.zshrc" ]; then
+    echo "Linking ZSH file"
+    ln -sf "$DOTFILES_DIR/shell/zsh/.zshrc" "$HOME/.zshrc"
+    echo "Installing OhMyZsh"
+    sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
+else
+    "No ZSHFILE found, killing the script ..."
+    exit
+fi
 echo "-----------------------------------------------"
 
 # Package Managers and packages
+
+# Install Linuxbrew
 echo "# Package managers & packages ✏️"
-/bin/bash "$DOTFILES_DIR/install/brew.sh"
-/bin/bash "$DOTFILES_DIR/install/volta.sh"
-git clone https://github.com/NvChad/NvChad ~/.config/nvim --depth 1
-sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
-/bin/bash "$DOTFILES_DIR/install/ssh.sh"
+if [ ! -d "/home/linuxbrew" ]; then
+    echo "Installing linuxbrew ..."
+    /bin/bash "$DOTFILES_DIR/install/brew.sh"
+    source ~/.zshrc
+fi
+
+# Install Volta
+if [ -d "/home/linuxbrew" ]; then
+    echo "Install Volta"
+    /bin/bash "$DOTFILES_DIR/install/volta.sh"
+fi
+# Nvchad
+if [ -d "/home/linuxbrew/.linuxbrew/Cellar/neovim" ]; then
+    echo "Install Nvchad"
+    git clone https://github.com/NvChad/NvChad ~/.config/nvim --depth 1
+fi
+
+# Generate SSH
+if [ ! -d "/home/dragutesku/.ssh" ]; then
+    echo "Generating new ssh"
+    /bin/bash "$DOTFILES_DIR/install/ssh.sh"
+fi
 echo "-----------------------------------------------"
 
 #Cleanup
